chainlit-service/requirements.txt
chainlit==1.0.200
openai==1.3.7
python-dotenv==1.0.0
redis==5.0.1
langroid==0.1.200
asyncio-mqtt==0.13.0
requests==2.31.0 